Refactor link widget stuff.
