+++ services/user/Chainmail/ui/src/components/editor/link-widget.tsx
@@ -1,136 +1,64 @@
-import { commandsCtx } from "@milkdown/core";
-import { updateLinkCommand } from "@milkdown/preset-commonmark";
-import { Plugin } from "@milkdown/prose/state";
-import { DecorationSet } from "@milkdown/prose/view";
-import { useInstance } from "@milkdown/react";
-import { $prose } from "@milkdown/utils";
-import type { useWidgetViewFactory } from "@prosemirror-adapter/react";
-import { useWidgetViewContext } from "@prosemirror-adapter/react";
+import type { Ctx } from "@milkdown/ctx";
+import type { EditorView } from "@milkdown/prose/view";
+import type { EditorState } from "@milkdown/prose/state";
 
-export const LinkWidgetBefore = () => {
-    return <>[</>;
-};
+import { editorViewCtx } from "@milkdown/core";
+import { linkSchema } from "@milkdown/preset-commonmark";
+import {
+    configureLinkTooltip,
+    linkTooltipAPI,
+    linkTooltipState,
+} from "@milkdown/components/link-tooltip";
+import { TooltipProvider, tooltipFactory } from "@milkdown/plugin-tooltip";
 
-export const LinkWidgetAfter = () => {
-    const { spec } = useWidgetViewContext();
-    const [loading, editor] = useInstance();
-    const href = spec?.href ?? "";
-    const title = spec?.title ?? "";
+import "../../styles/link-tooltip.css";
 
-    return (
-        <>
-            ]
-            <span>
-                (
-                {
-                    <>
-                        <small className="text-nord8 font-light">link: </small>
-                        <input
-                            size={href.length}
-                            placeholder="empty"
-                            onBlur={(e) => {
-                                if (loading) return;
-                                editor().action((ctx) => {
-                                    const commands = ctx.get(commandsCtx);
-                                    commands.call(updateLinkCommand.key, {
-                                        href: e.target.value,
-                                    });
-                                });
-                            }}
-                            className="rounded border-none bg-gray-50 px-2 py-0 ring-1 dark:bg-gray-900"
-                            type="text"
-                            defaultValue={href}
-                        />
-                        &nbsp;
-                        <small className="text-nord8 font-light">title: </small>
-                        &quot;
-                        <input
-                            size={title.length || 5}
-                            placeholder="empty"
-                            onBlur={(e) => {
-                                if (loading) return;
-                                editor().action((ctx) => {
-                                    const commands = ctx.get(commandsCtx);
-                                    commands.call(updateLinkCommand.key, {
-                                        title: e.target.value,
-                                    });
-                                });
-                            }}
-                            className="rounded border-none bg-gray-50 px-2 py-0 ring-1 dark:bg-gray-900"
-                            type="text"
-                            defaultValue={title}
-                        />
-                        &quot;
-                    </>
-                }
-                )
-            </span>
-        </>
-    );
-};
+export { linkTooltipPlugin } from "@milkdown/components/link-tooltip";
 
-export const linkPlugin = (
-    widgetViewFactory: ReturnType<typeof useWidgetViewFactory>,
-) => {
-    const before = widgetViewFactory({
-        as: "span",
-        component: LinkWidgetBefore,
+export const insertLinkTooltip = tooltipFactory("CREATE_LINK");
+
+export function tooltipPluginView(ctx: Ctx) {
+    return (_view: EditorView) => {
+        const content = document.createElement("div");
+        const provider = new TooltipProvider({
+            content,
+            shouldShow: (view: EditorView) => {
+                const { selection, doc } = view.state;
+                const has = doc.rangeHasMark(
+                    selection.from,
+                    selection.to,
+                    linkSchema.type(ctx),
+                );
+                if (has || selection.empty) return false;
+
+                return true;
+            },
+        });
+
+        content.onmousedown = (e: MouseEvent) => {
+            e.preventDefault();
+            const view = ctx.get(editorViewCtx);
+            const { selection } = view.state;
+            ctx.get(linkTooltipAPI.key).addLink(selection.from, selection.to);
+            provider.hide();
+        };
+
+        return {
+            update: (updatedView: EditorView, prevState: EditorState) => {
+                if (ctx.get(linkTooltipState.key).mode === "edit") return;
+                provider.update(updatedView, prevState);
+            },
+            destroy: () => {
+                provider.destroy();
+                content.remove();
+            },
+        };
+    };
+}
+
+export function linkTooltipConfig(ctx: Ctx) {
+    ctx.set(insertLinkTooltip.key, {
+        view: tooltipPluginView(ctx),
     });
-    const after = widgetViewFactory({ as: "span", component: LinkWidgetAfter });
-
-    return $prose(
-        () =>
-            new Plugin({
-                state: {
-                    init() {
-                        return DecorationSet.empty;
-                    },
-                    apply(tr) {
-                        const { selection } = tr;
-
-                        const { $from, $to } = selection;
-                        const node = tr.doc.nodeAt(selection.from);
-
-                        const mark = node?.marks.find(
-                            (mark) => mark.type.name === "link",
-                        );
-
-                        if (!mark) return DecorationSet.empty;
-
-                        let markPos = { start: -1, end: -1 };
-                        tr.doc.nodesBetween(
-                            $from.start(),
-                            $to.end(),
-                            (n, pos) => {
-                                if (node === n) {
-                                    markPos = {
-                                        start: pos,
-                                        end:
-                                            pos +
-                                            Math.max(n.textContent.length, 1),
-                                    };
-
-                                    // stop recursing if result is found
-                                    return false;
-                                }
-                                return undefined;
-                            },
-                        );
-
-                        return DecorationSet.create(tr.doc, [
-                            before(markPos.start),
-                            after(markPos.end, {
-                                href: mark.attrs.href,
-                                title: mark.attrs.title,
-                            }),
-                        ]);
-                    },
-                },
-                props: {
-                    decorations(state) {
-                        return this.getState(state);
-                    },
-                },
-            }),
-    );
-};
+    configureLinkTooltip(ctx);
+}
